Convert messages to TwitchMessage on load
This will be a bit slower, but in most cases the waiting client(s) (if
any) will only need the data of the first load anyway to continue.

The upside of this is that this has a significantly reduced memory
footprint (by roughly tenfold, although part of this will be lost once
TwitchMessage is extended with things like colors and emotes).

Another added benefit is that TwitchChat now no longer needs to know
anything about the format of a twitch message.